Remove dependency on mscorlib from portable C# script tests
