+++ src/Scripting/CSharpTest/InteractiveSessionTests.cs
@@ -30,12 +30,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Scripting.UnitTests
 
     public class InteractiveSessionTests : TestBase
     {
-        private static readonly CSharpCompilationOptions s_signedDll =
-            new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary, cryptoPublicKey: TestResources.TestKeys.PublicKey_ce65828c82a341f2);
-
-        private static readonly CSharpCompilationOptions s_signedDll2 =
-            new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary, cryptoPublicKey: TestResources.TestKeys.PublicKey_ce65828c82a341f2);
-
         internal static readonly Assembly HostAssembly = typeof(InteractiveSessionTests).GetTypeInfo().Assembly;
 
         #region Namespaces, Types
@@ -1134,14 +1128,29 @@ static T G<T>(T t, Func<T, Task<T>> f)
         [Fact]
         public void ReferenceDirective_FileWithDependencies()
         {
-            string file1 = Temp.CreateFile().WriteAllBytes(TestResources.MetadataTests.InterfaceAndClass.CSClasses01).Path;
-            string file2 = Temp.CreateFile().WriteAllBytes(TestResources.MetadataTests.InterfaceAndClass.CSInterfaces01).Path;
+            var file1 = Temp.CreateFile();
+            var file2 = Temp.CreateFile();
 
-            // ICSPropImpl in CSClasses01.dll implements ICSProp in CSInterfces01.dll.
-            object result = CSharpScript.EvaluateAsync(@"
-#r """ + file1 + @"""
-#r """ + file2 + @"""
-new Metadata.ICSPropImpl()
+            var lib1 = CreateCSharpCompilationWithCorlib(@"
+public interface I
+{
+    int F();
+}");
+
+            lib1.Emit(file1.Path);
+
+            var lib2 = CreateCSharpCompilation(@" 
+public class C : I
+{
+    public int F() => 1;
+}", new MetadataReference[] { TestReferences.NetStandard13.SystemRuntime, lib1.ToMetadataReference() });
+
+            lib2.Emit(file2.Path);
+
+            object result = CSharpScript.EvaluateAsync($@"
+#r ""{file1.Path}""
+#r ""{file2.Path}""
+new C()
 ").Result;
             Assert.NotNull(result);
         }
@@ -1149,12 +1158,17 @@ new Metadata.ICSPropImpl()
         [Fact]
         public void ReferenceDirective_RelativeToBaseParent()
         {
-            string path = Temp.CreateFile().WriteAllBytes(TestResources.MetadataTests.InterfaceAndClass.CSClasses01).Path;
-            string fileName = Path.GetFileName(path);
-            string dir = Path.Combine(Path.GetDirectoryName(path), "subdir");
+            var file = Temp.CreateFile();
+            var lib = CreateCSharpCompilationWithCorlib("public class C {}");
+            lib.Emit(file.Path);
 
-            var script = CSharpScript.Create($@"#r ""..\{fileName}""",
-                ScriptOptions.Default.WithFilePath(Path.Combine(dir, "a.csx")));
+            string dir = Path.Combine(Path.GetDirectoryName(file.Path), "subdir");
+            string libFileName = Path.GetFileName(file.Path);
+            string scriptPath = Path.Combine(dir, "a.csx");
+
+            var script = CSharpScript.Create(
+                $@"#r ""..\{libFileName}""",
+                ScriptOptions.Default.WithFilePath(scriptPath));
 
             script.GetCompilation().VerifyDiagnostics();
         }
@@ -1162,551 +1176,21 @@ new Metadata.ICSPropImpl()
         [Fact]
         public void ReferenceDirective_RelativeToBaseRoot()
         {
-            string path = Temp.CreateFile().WriteAllBytes(TestResources.MetadataTests.InterfaceAndClass.CSClasses01).Path;
-            string root = Path.GetPathRoot(path);
-            string unrooted = path.Substring(root.Length);
+            var file = Temp.CreateFile();
+            var lib = CreateCSharpCompilationWithCorlib("public class C {}");
+            lib.Emit(file.Path);
+
+            string root = Path.GetPathRoot(file.Path);
+            string unrooted = file.Path.Substring(root.Length);
 
             string dir = Path.Combine(root, "foo", "bar", "baz");
+            string scriptPath = Path.Combine(dir, "a.csx");
 
-            var script = CSharpScript.Create($@"#r ""\{unrooted}""",
-                ScriptOptions.Default.WithFilePath(Path.Combine(dir, "a.csx")));
+            var script = CSharpScript.Create(
+                $@"#r ""\{unrooted}""",
+                ScriptOptions.Default.WithFilePath(scriptPath));
 
             script.GetCompilation().VerifyDiagnostics();
-        }
-
-        [Fact, WorkItem(6457, "https://github.com/dotnet/roslyn/issues/6457")]
-        public async Task MissingReferencesReuse()
-        {
-            var source = @"
-public class C
-{
-    public System.Diagnostics.Process P;
-}
-";
-
-            var lib = CSharpCompilation.Create(
-                "Lib",
-                new[] { SyntaxFactory.ParseSyntaxTree(source) },
-                new[] { TestReferences.NetFx.v4_0_30319.mscorlib, TestReferences.NetFx.v4_0_30319.System },
-                new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary));
-
-            var libFile = Temp.CreateFile("lib").WriteAllBytes(lib.EmitToArray());
-
-            var s0 = await CSharpScript.RunAsync("C c;", ScriptOptions.Default.WithReferences(libFile.Path));
-            var s1 = await s0.ContinueWithAsync("c = new C()");
-        }
-
-        [Fact]
-        public async Task SharedLibCopy_Identical_Weak()
-        {
-            string libBaseName = "LibBase_" + Guid.NewGuid();
-            string lib1Name = "Lib1_" + Guid.NewGuid();
-            string lib2Name = "Lib2_" + Guid.NewGuid();
-
-            var libBase = CreateCSharpCompilation(@"
-public class LibBase
-{
-    public readonly int X = 1;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName);
-
-            var lib1 = CreateCSharpCompilation(@"
-public class Lib1
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase.ToMetadataReference() }, lib1Name);
-
-            var lib2 = CreateCSharpCompilation(@"
-public class Lib2
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase.ToMetadataReference() }, lib2Name);
-
-            var libBaseImage = libBase.EmitToArray();
-            var lib1Image = lib1.EmitToArray();
-            var lib2Image = lib2.EmitToArray();
-
-            var root = Temp.CreateDirectory();
-            var dir1 = root.CreateDirectory("1");
-            var file1 = dir1.CreateFile(lib1Name + ".dll").WriteAllBytes(lib1Image);
-            var fileBase1 = dir1.CreateFile(libBaseName + ".dll").WriteAllBytes(libBaseImage);
-
-            var dir2 = root.CreateDirectory("2");
-            var file2 = dir2.CreateFile(lib2Name + ".dll").WriteAllBytes(lib2Image);
-            var fileBase2 = dir2.CreateFile(libBaseName + ".dll").WriteAllBytes(libBaseImage);
-
-            var s0 = await CSharpScript.RunAsync($@"#r ""{file1.Path}""");
-            var s1 = await s0.ContinueWithAsync($@"var l1 = new Lib1();");
-            var s2 = await s1.ContinueWithAsync($@"#r ""{file2.Path}""");
-            var s3 = await s2.ContinueWithAsync($@"var l2 = new Lib2();");
-            var s4 = await s3.ContinueWithAsync($@"l2.libBase.X");
-
-            var c4 = s4.Script.GetCompilation();
-            c4.VerifyAssemblyAliases(
-                lib2Name,
-                lib1Name,
-                "mscorlib",
-                libBaseName + ": <implicit>,global");
-
-            var libBaseRefAndSymbol = c4.GetBoundReferenceManager().GetReferencedAssemblies().ToArray()[3];
-            Assert.Equal(fileBase1.Path, ((PortableExecutableReference)libBaseRefAndSymbol.Key).FilePath);
-        }
-
-        [Fact]
-        public async Task SharedLibCopy_Identical_Strong()
-        {
-            string libBaseName = "LibBase_" + Guid.NewGuid();
-            string lib1Name = "Lib1_" + Guid.NewGuid();
-            string lib2Name = "Lib2_" + Guid.NewGuid();
-
-            var libBase = CreateCSharpCompilation(@"
-public class LibBase
-{
-    public readonly int X = 1;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName, s_signedDll);
-
-            var lib1 = CreateCSharpCompilation(@"
-public class Lib1
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase.ToMetadataReference() }, lib1Name);
-
-            var lib2 = CreateCSharpCompilation(@"
-public class Lib2
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase.ToMetadataReference() }, lib2Name);
-
-            var libBaseImage = libBase.EmitToArray();
-            var lib1Image = lib1.EmitToArray();
-            var lib2Image = lib2.EmitToArray();
-
-            var root = Temp.CreateDirectory();
-            var dir1 = root.CreateDirectory("1");
-            var file1 = dir1.CreateFile(lib1Name + ".dll").WriteAllBytes(lib1Image);
-            var fileBase1 = dir1.CreateFile(libBaseName + ".dll").WriteAllBytes(libBaseImage);
-
-            var dir2 = root.CreateDirectory("2");
-            var file2 = dir2.CreateFile(lib2Name + ".dll").WriteAllBytes(lib2Image);
-            var fileBase2 = dir2.CreateFile(libBaseName + ".dll").WriteAllBytes(libBaseImage);
-
-            var s0 = await CSharpScript.RunAsync($@"#r ""{file1.Path}""");
-            var s1 = await s0.ContinueWithAsync($@"var l1 = new Lib1();");
-            var s2 = await s1.ContinueWithAsync($@"#r ""{file2.Path}""");
-            var s3 = await s2.ContinueWithAsync($@"var l2 = new Lib2();");
-            var s4 = await s3.ContinueWithAsync($@"l2.libBase.X");
-
-            var c4 = s4.Script.GetCompilation();
-            c4.VerifyAssemblyAliases(
-                lib2Name,
-                lib1Name,
-                "mscorlib",
-                libBaseName + ": <implicit>,global");
-
-            var libBaseRefAndSymbol = c4.GetBoundReferenceManager().GetReferencedAssemblies().ToArray()[3];
-            Assert.Equal(fileBase1.Path, ((PortableExecutableReference)libBaseRefAndSymbol.Key).FilePath);
-        }
-
-        [Fact]
-        public async Task SharedLibCopy_SameVersion_Weak_DifferentContent()
-        {
-            string libBaseName = "LibBase_" + Guid.NewGuid();
-            string lib1Name = "Lib1_" + Guid.NewGuid();
-            string lib2Name = "Lib2_" + Guid.NewGuid();
-
-            var libBase1 = CreateCSharpCompilation(@"
-public class LibBase
-{
-    public readonly int X = 1;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName);
-
-            var libBase2 = CreateCSharpCompilation(@"
-public class LibBase
-{
-    public readonly int X = 2;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName);
-
-            var lib1 = CreateCSharpCompilation(@"
-public class Lib1
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib1Name);
-
-            var lib2 = CreateCSharpCompilation(@"
-public class Lib2
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib2Name);
-
-            var libBase1Image = libBase1.EmitToArray();
-            var libBase2Image = libBase2.EmitToArray();
-            var lib1Image = lib1.EmitToArray();
-            var lib2Image = lib2.EmitToArray();
-
-            var root = Temp.CreateDirectory();
-            var dir1 = root.CreateDirectory("1");
-            var file1 = dir1.CreateFile(lib1Name + ".dll").WriteAllBytes(lib1Image);
-            var fileBase1 = dir1.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase1Image);
-
-            var dir2 = root.CreateDirectory("2");
-            var file2 = dir2.CreateFile(lib2Name + ".dll").WriteAllBytes(lib2Image);
-            var fileBase2 = dir2.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase2Image);
-
-            var s0 = await CSharpScript.RunAsync($@"#r ""{file1.Path}""");
-            var s1 = await s0.ContinueWithAsync($@"var l1 = new Lib1();");
-            var s2 = await s1.ContinueWithAsync($@"#r ""{file2.Path}""");
-
-            bool exceptionSeen = false;
-            try
-            {
-                await s2.ContinueWithAsync($@"var l2 = new Lib2();");
-            }
-            catch (FileLoadException fileLoadEx) when (fileLoadEx.InnerException is InteractiveAssemblyLoaderException)
-            {
-                exceptionSeen = true;
-            }
-
-            Assert.True(exceptionSeen);
-        }
-
-        [Fact]
-        public async Task SharedLibCopy_SameVersion_Strong_DifferentContent()
-        {
-            string libBaseName = "LibBase_" + Guid.NewGuid();
-            string lib1Name = "Lib1_" + Guid.NewGuid();
-            string lib2Name = "Lib2_" + Guid.NewGuid();
-
-            var libBase1 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""1.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 1;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName, s_signedDll);
-
-            var libBase2 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""1.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 2;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName, s_signedDll);
-
-            var lib1 = CreateCSharpCompilation(@"
-public class Lib1
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib1Name);
-
-            var lib2 = CreateCSharpCompilation(@"
-public class Lib2
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib2Name);
-
-            var libBase1Image = libBase1.EmitToArray();
-            var libBase2Image = libBase2.EmitToArray();
-            var lib1Image = lib1.EmitToArray();
-            var lib2Image = lib2.EmitToArray();
-
-            var root = Temp.CreateDirectory();
-            var dir1 = root.CreateDirectory("1");
-            var file1 = dir1.CreateFile(lib1Name + ".dll").WriteAllBytes(lib1Image);
-            var fileBase1 = dir1.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase1Image);
-
-            var dir2 = root.CreateDirectory("2");
-            var file2 = dir2.CreateFile(lib2Name + ".dll").WriteAllBytes(lib2Image);
-            var fileBase2 = dir2.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase2Image);
-
-            var s0 = await CSharpScript.RunAsync($@"#r ""{file1.Path}""");
-            var s1 = await s0.ContinueWithAsync($@"new Lib1().libBase.X");
-            var s2 = await s1.ContinueWithAsync($@"#r ""{file2.Path}""");
-
-            bool exceptionSeen = false;
-            try
-            {
-                await s2.ContinueWithAsync($@"new Lib2().libBase.X");
-            }
-            catch (FileLoadException fileLoadEx) when (fileLoadEx.InnerException is InteractiveAssemblyLoaderException)
-            {
-                exceptionSeen = true;
-            }
-
-            Assert.True(exceptionSeen);
-        }
-
-        [Fact]
-        public async Task SharedLibCopy_SameVersion_StrongWeak_DifferentContent()
-        {
-            string libBaseName = "LibBase_" + Guid.NewGuid();
-            string lib1Name = "Lib1_" + Guid.NewGuid();
-            string lib2Name = "Lib2_" + Guid.NewGuid();
-
-            var libBase1 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""1.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 1;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName, s_signedDll);
-
-            var libBase2 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""1.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 2;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName);
-
-            var lib1 = CreateCSharpCompilation(@"
-public class Lib1
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib1Name);
-
-            var lib2 = CreateCSharpCompilation(@"
-public class Lib2
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib2Name);
-
-            var libBase1Image = libBase1.EmitToArray();
-            var libBase2Image = libBase2.EmitToArray();
-            var lib1Image = lib1.EmitToArray();
-            var lib2Image = lib2.EmitToArray();
-
-            var root = Temp.CreateDirectory();
-            var dir1 = root.CreateDirectory("1");
-            var file1 = dir1.CreateFile(lib1Name + ".dll").WriteAllBytes(lib1Image);
-            var fileBase1 = dir1.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase1Image);
-
-            var dir2 = root.CreateDirectory("2");
-            var file2 = dir2.CreateFile(lib2Name + ".dll").WriteAllBytes(lib2Image);
-            var fileBase2 = dir2.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase2Image);
-
-            var s0 = await CSharpScript.RunAsync($@"#r ""{file1.Path}""");
-            var s1 = await s0.ContinueWithAsync($@"new Lib1().libBase.X");
-            var s2 = await s1.ContinueWithAsync($@"#r ""{file2.Path}""");
-
-            bool exceptionSeen = false;
-            try
-            {
-                await s2.ContinueWithAsync($@"new Lib2().libBase.X");
-            }
-            catch (FileLoadException fileLoadEx) when (fileLoadEx.InnerException is InteractiveAssemblyLoaderException)
-            {
-                exceptionSeen = true;
-            }
-
-            Assert.True(exceptionSeen);
-        }
-
-        [Fact]
-        public async Task SharedLibCopy_SameVersion_StrongDifferentPKT_DifferentContent()
-        {
-            string libBaseName = "LibBase_" + Guid.NewGuid();
-            string lib1Name = "Lib1_" + Guid.NewGuid();
-            string lib2Name = "Lib2_" + Guid.NewGuid();
-
-            var libBase1 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""1.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 1;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName, s_signedDll);
-
-            var libBase2 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""1.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 2;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName, s_signedDll2);
-
-            var lib1 = CreateCSharpCompilation(@"
-public class Lib1
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib1Name);
-
-            var lib2 = CreateCSharpCompilation(@"
-public class Lib2
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib2Name);
-
-            var libBase1Image = libBase1.EmitToArray();
-            var libBase2Image = libBase2.EmitToArray();
-            var lib1Image = lib1.EmitToArray();
-            var lib2Image = lib2.EmitToArray();
-
-            var root = Temp.CreateDirectory();
-            var dir1 = root.CreateDirectory("1");
-            var file1 = dir1.CreateFile(lib1Name + ".dll").WriteAllBytes(lib1Image);
-            var fileBase1 = dir1.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase1Image);
-
-            var dir2 = root.CreateDirectory("2");
-            var file2 = dir2.CreateFile(lib2Name + ".dll").WriteAllBytes(lib2Image);
-            var fileBase2 = dir2.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase2Image);
-
-            var s0 = await CSharpScript.RunAsync($@"#r ""{file1.Path}""");
-            var s1 = await s0.ContinueWithAsync($@"new Lib1().libBase.X");
-            var s2 = await s1.ContinueWithAsync($@"#r ""{file2.Path}""");
-
-            bool exceptionSeen = false;
-            try
-            {
-                await s2.ContinueWithAsync($@"new Lib2().libBase.X");
-            }
-            catch (FileLoadException fileLoadEx) when (fileLoadEx.InnerException is InteractiveAssemblyLoaderException)
-            {
-                exceptionSeen = true;
-            }
-
-            Assert.True(exceptionSeen);
-        }
-
-        [Fact]
-        public async Task SharedLibCopy_DifferentVersion_Weak()
-        {
-            string libBaseName = "LibBase_" + Guid.NewGuid();
-            string lib1Name = "Lib1_" + Guid.NewGuid();
-            string lib2Name = "Lib2_" + Guid.NewGuid();
-
-            var libBase1 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""1.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 1;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName);
-
-            var libBase2 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""2.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 2;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName);
-
-            var lib1 = CreateCSharpCompilation(@"
-public class Lib1
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib1Name);
-
-            var lib2 = CreateCSharpCompilation(@"
-public class Lib2
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase2.ToMetadataReference() }, lib2Name);
-
-            var libBase1Image = libBase1.EmitToArray();
-            var libBase2Image = libBase2.EmitToArray();
-            var lib1Image = lib1.EmitToArray();
-            var lib2Image = lib2.EmitToArray();
-
-            var root = Temp.CreateDirectory();
-            var dir1 = root.CreateDirectory("1");
-            var file1 = dir1.CreateFile(lib1Name + ".dll").WriteAllBytes(lib1Image);
-            var fileBase1 = dir1.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase1Image);
-
-            var dir2 = root.CreateDirectory("2");
-            var file2 = dir2.CreateFile(lib2Name + ".dll").WriteAllBytes(lib2Image);
-            var fileBase2 = dir2.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase2Image);
-
-            var s0 = await CSharpScript.RunAsync($@"#r ""{file1.Path}""");
-            var s1 = await s0.ContinueWithAsync($@"var l1 = new Lib1().libBase.X;");
-            var s2 = await s1.ContinueWithAsync($@"#r ""{file2.Path}""");
-
-            bool exceptionSeen = false;
-            try
-            {
-                await s2.ContinueWithAsync($@"var l2 = new Lib2().libBase.X;");
-            }
-            catch (FileLoadException fileLoadEx) when (fileLoadEx.InnerException is InteractiveAssemblyLoaderException)
-            {
-                exceptionSeen = true;
-            }
-
-            Assert.True(exceptionSeen);
-        }
-
-        [Fact]
-        public async Task SharedLibCopy_DifferentVersion_Strong()
-        {
-            string libBaseName = "LibBase_" + Guid.NewGuid();
-            string lib1Name = "Lib1_" + Guid.NewGuid();
-            string lib2Name = "Lib2_" + Guid.NewGuid();
-
-            var libBase1 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""1.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 1;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName, s_signedDll);
-
-            var libBase2 = CreateCSharpCompilation(@"
-[assembly: System.Reflection.AssemblyVersion(""2.0.0.0"")]
-public class LibBase
-{
-    public readonly int X = 2;
-}
-", new[] { TestReferences.NetFx.v4_0_30319.mscorlib }, libBaseName, s_signedDll);
-
-            var lib1 = CreateCSharpCompilation(@"
-public class Lib1
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase1.ToMetadataReference() }, lib1Name);
-
-            var lib2 = CreateCSharpCompilation(@"
-public class Lib2
-{
-    public LibBase libBase = new LibBase();
-}
-", new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libBase2.ToMetadataReference() }, lib2Name);
-
-            var libBase1Image = libBase1.EmitToArray();
-            var libBase2Image = libBase2.EmitToArray();
-            var lib1Image = lib1.EmitToArray();
-            var lib2Image = lib2.EmitToArray();
-
-            var root = Temp.CreateDirectory();
-            var dir1 = root.CreateDirectory("1");
-            var file1 = dir1.CreateFile(lib1Name + ".dll").WriteAllBytes(lib1Image);
-            var fileBase1 = dir1.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase1Image);
-
-            var dir2 = root.CreateDirectory("2");
-            var file2 = dir2.CreateFile(lib2Name + ".dll").WriteAllBytes(lib2Image);
-            var fileBase2 = dir2.CreateFile(libBaseName + ".dll").WriteAllBytes(libBase2Image);
-
-            var s0 = await CSharpScript.RunAsync($@"#r ""{file1.Path}""");
-            var s1 = await s0.ContinueWithAsync($@"new Lib1().libBase.X");
-            Assert.Equal(1, s1.ReturnValue);
-            var s2 = await s1.ContinueWithAsync($@"#r ""{file2.Path}""");
-            var s3 = await s2.ContinueWithAsync($@"new Lib2().libBase.X");
-            Assert.Equal(2, s3.ReturnValue);
         }
 
         [Fact]
@@ -1715,13 +1199,13 @@ public class Lib2
             string mainName = "Main_" + Guid.NewGuid();
             string libName = "Lib_" + Guid.NewGuid();
 
-            var libExe = TestCompilationFactory.CreateCSharpCompilationWithMscorlib(@"public class C { public string F = ""exe""; }", libName);
-            var libDll = TestCompilationFactory.CreateCSharpCompilationWithMscorlib(@"public class C { public string F = ""dll""; }", libName);
-            var libWinmd = TestCompilationFactory.CreateCSharpCompilationWithMscorlib(@"public class C { public string F = ""winmd""; }", libName);
+            var libExe = CreateCSharpCompilationWithCorlib(@"public class C { public string F = ""exe""; }", libName);
+            var libDll = CreateCSharpCompilationWithCorlib(@"public class C { public string F = ""dll""; }", libName);
+            var libWinmd = CreateCSharpCompilationWithCorlib(@"public class C { public string F = ""winmd""; }", libName);
 
             var main = CreateCSharpCompilation(
                 @"public static class M { public static readonly C X = new C(); }",
-                new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libExe.ToMetadataReference() },
+                new MetadataReference[] { TestReferences.NetStandard13.SystemRuntime, libExe.ToMetadataReference() },
                 mainName);
 
             var exeImage = libExe.EmitToArray();
@@ -1745,13 +1229,13 @@ public class Lib2
             string mainName = "Main_" + Guid.NewGuid();
             string libName = "Lib_" + Guid.NewGuid();
 
-            var libExe = TestCompilationFactory.CreateCSharpCompilationWithMscorlib(@"public class C { public string F = ""exe""; }", libName);
-            var libDll = TestCompilationFactory.CreateCSharpCompilationWithMscorlib(@"public class C { public string F = ""dll""; }", libName);
-            var libWinmd = TestCompilationFactory.CreateCSharpCompilationWithMscorlib(@"public class C { public string F = ""winmd""; }", libName);
+            var libExe = CreateCSharpCompilationWithCorlib(@"public class C { public string F = ""exe""; }", libName);
+            var libDll = CreateCSharpCompilationWithCorlib(@"public class C { public string F = ""dll""; }", libName);
+            var libWinmd = CreateCSharpCompilationWithCorlib(@"public class C { public string F = ""winmd""; }", libName);
 
             var main = CreateCSharpCompilation(
                 @"public static class M { public static readonly C X = new C(); }",
-                new MetadataReference[] { TestReferences.NetFx.v4_0_30319.mscorlib, libExe.ToMetadataReference() },
+                new MetadataReference[] { TestReferences.NetStandard13.SystemRuntime, libExe.ToMetadataReference() },
                 mainName);
 
             var exeImage = libExe.EmitToArray();
@@ -1779,7 +1263,7 @@ public class D { }
 public class E { }
 ";
 
-            var libRef = CreateCSharpCompilationWithMscorlib(source, "lib").EmitToImageReference();
+            var libRef = CreateCSharpCompilationWithCorlib(source, "lib").EmitToImageReference();
 
             var script = CSharpScript.Create(@"new C()",
                 ScriptOptions.Default.WithReferences(libRef.WithAliases(new[] { "Hidden" })).WithImports("Hidden::N"));
